requirements.txt
crowdstrike-falconpy>=1.4.0
python-dotenv>=1.0.0
requests>=2.31.0
tabulate>=0.9.0
colorama>=0.4.6
tqdm>=4.66.0
